# distribution/resources/postman/Beer_Control.swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"REST API for beer stock management",
    "version":"1.0.0",
    "title":"Beer Control API",
    "contact":{
      "name":"Samuel Cunha",
      "url":"https://gtihub.com/sawcunha",
      "email":"samukasaww@live.com"
    }
  },
  "host":"localhost:8080",
  "basePath":"/beercontrol",
...
  "paths":{
...
    "/api/v1/address/person/{idPerson}":{
      "get":{
        "tags":[
          "address-controller"
        ],
        "summary":"findByPersonId",
        "operationId":"findByPersonIdUsingGET",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"idPerson",
            "in":"path",
            "description":"idPerson",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«List«AddressResponseDTO»»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "post":{
        "tags":[
          "address-controller"
        ],
        "summary":"createAddressPerson",
        "operationId":"createAddressPersonUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"addressRequestDTO",
            "description":"addressRequestDTO",
            "required":true,
            "schema":{
              "$ref":"#/definitions/AddressRequestDTO"
            }
          },
          {
            "name":"idPerson",
            "in":"path",
            "description":"idPerson",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "201":{
            "description":"Created",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«AddressResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      }
    },
    "/api/v1/address/{id}":{
      "get":{
        "tags":[
          "address-controller"
        ],
        "summary":"findById",
        "operationId":"findByIdUsingGET",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          },
          {
            "name":"person",
            "in":"query",
            "description":"person",
            "required":false,
            "type":"boolean",
            "default":false
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«AddressResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "put":{
        "tags":[
          "address-controller"
        ],
        "summary":"updateAddressPerson",
        "operationId":"updateAddressPersonUsingPUT",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"addressRequestDTO",
            "description":"addressRequestDTO",
            "required":true,
            "schema":{
              "$ref":"#/definitions/AddressRequestDTO"
            }
          },
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«AddressResponseDTO»"
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "address-controller"
        ],
        "summary":"delete",
        "operationId":"deleteUsingDELETE",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    },
...
    "/api/v1/beer/{id}":{
      "get":{
        "tags":[
          "beer-controller"
        ],
        "summary":"findById",
        "operationId":"findByIdUsingGET_1",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«BeerResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "beer-controller"
        ],
        "summary":"delete",
        "operationId":"deleteUsingDELETE_1",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "202":{
            "description":"Accepted"
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    },
...
    "/api/v1/person":{
      "get":{
        "tags":[
          "person-controller"
        ],
        "summary":"findAll",
        "operationId":"findAllUsingGET_2",
        "produces":[
          "application/json"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«List«PersonResponseDTO»»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "post":{
        "tags":[
          "person-controller"
        ],
        "summary":"save",
        "operationId":"saveUsingPOST_2",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"personDTO",
            "description":"personDTO",
            "required":true,
            "schema":{
              "$ref":"#/definitions/PersonRequestDTO"
            }
          }
        ],
        "responses":{
          "201":{
            "description":"Created",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«PersonResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      }
    },
...
    "/api/v1/person/{id}":{
      "get":{
        "tags":[
          "person-controller"
        ],
        "summary":"findById",
        "operationId":"findByIdUsingGET_2",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«PersonResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "put":{
        "tags":[
          "person-controller"
        ],
        "summary":"update",
        "operationId":"updateUsingPUT_2",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          },
          {
            "in":"body",
            "name":"personDTO",
            "description":"personDTO",
            "required":true,
            "schema":{
              "$ref":"#/definitions/PersonRequestDTO"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«PersonResponseDTO»"
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "person-controller"
        ],
        "summary":"delete",
        "operationId":"deleteUsingDELETE_3",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    },
...
    "/api/v1/phone/person/{idPerson}":{
      "get":{
        "tags":[
          "phone-controller"
        ],
        "summary":"findByPersonId",
        "operationId":"findByPersonIdUsingGET_1",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"idPerson",
            "in":"path",
            "description":"idPerson",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«List«PhoneResponseDTO»»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "post":{
        "tags":[
          "phone-controller"
        ],
        "summary":"createPhonePerson",
        "operationId":"createPhonePersonUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"idPerson",
            "in":"path",
            "description":"idPerson",
            "required":true,
            "type":"integer",
            "format":"int64"
          },
          {
            "in":"body",
            "name":"phoneDTO",
            "description":"phoneDTO",
            "required":true,
            "schema":{
              "$ref":"#/definitions/PhoneRequestDTO"
            }
          }
        ],
        "responses":{
          "201":{
            "description":"Created",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«PhoneResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      }
    },
    "/api/v1/phone/{id}":{
      "get":{
        "tags":[
          "phone-controller"
        ],
        "summary":"findById",
        "operationId":"findByIdUsingGET_3",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          },
          {
            "name":"person",
            "in":"query",
            "description":"person",
            "required":false,
            "type":"boolean",
            "default":false
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«PhoneResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "put":{
        "tags":[
          "phone-controller"
        ],
        "summary":"updatePhonePerson",
        "operationId":"updatePhonePersonUsingPUT",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          },
          {
            "in":"body",
            "name":"phoneDTO",
            "description":"phoneDTO",
            "required":true,
            "schema":{
              "$ref":"#/definitions/PhoneRequestDTO"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«PhoneResponseDTO»"
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "phone-controller"
        ],
        "summary":"delete",
        "operationId":"deleteUsingDELETE_4",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    },
...
    "/api/v1/warehouse/{id}":{
      "get":{
        "tags":[
          "warehouse-controller"
        ],
        "summary":"findById",
        "operationId":"findByIdUsingGET_4",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/BeerControlResponse«WarehouseResponseDTO»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "warehouse-controller"
        ],
        "summary":"delete",
        "operationId":"deleteUsingDELETE_5",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    }
  },
  "definitions":{
    "AddressPersonDTO":{
      "type":"object",
      "properties":{
        "birthday":{
          "type":"string",
          "format":"date"
        },
        "cpf":{
          "type":"string"
        },
        "description":{
          "type":"string"
        },
        "email":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        },
        "sex":{
          "type":"string",
          "enum":[
            "ASEXUAL",
            "BISEXUAL",
            "FEMALE",
            "HOMOSEXUAL",
            "MALE",
            "PANSEXUAL"
          ]
        },
        "surname":{
          "type":"string"
        }
      },
      "title":"AddressPersonDTO"
    },
    "AddressPersonDTO0":{
      "type":"object",
      "properties":{
        "birthday":{
          "type":"string",
          "format":"date"
        },
        "cpf":{
          "type":"string"
        },
        "description":{
          "type":"string"
        },
        "email":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        },
        "sex":{
          "type":"string",
          "enum":[
            "ASEXUAL",
            "BISEXUAL",
            "FEMALE",
            "HOMOSEXUAL",
            "MALE",
            "PANSEXUAL"
          ]
        },
        "surname":{
          "type":"string"
        }
      },
      "title":"AddressPersonDTO0"
    },
    "AddressRequestDTO":{
      "type":"object",
      "properties":{
        "addressName":{
          "type":"string"
        },
        "block":{
          "type":"string"
        },
        "city":{
          "type":"string"
        },
        "complement":{
          "type":"string"
        },
        "country":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "number":{
          "type":"integer",
          "format":"int32"
        },
        "person":{
          "$ref":"#/definitions/AddressPersonDTO"
        },
        "reference":{
          "type":"string"
        },
        "state":{
          "type":"string"
        },
        "street":{
          "type":"string"
        },
        "zipCode":{
          "type":"string"
        }
      },
      "title":"AddressRequestDTO"
    },
    "AddressResponseDTO":{
      "type":"object",
      "properties":{
        "addressName":{
          "type":"string"
        },
        "block":{
          "type":"string"
        },
        "city":{
          "type":"string"
        },
        "complement":{
          "type":"string"
        },
        "country":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "number":{
          "type":"integer",
          "format":"int32"
        },
        "person":{
          "$ref":"#/definitions/AddressPersonDTO0"
        },
        "reference":{
          "type":"string"
        },
        "state":{
          "type":"string"
        },
        "street":{
          "type":"string"
        },
        "zipCode":{
          "type":"string"
        }
      },
      "title":"AddressResponseDTO"
    },
...
    "BeerControlResponse«AddressResponseDTO»":{
      "type":"object",
      "properties":{
        "data":{
          "$ref":"#/definitions/AddressResponseDTO"
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«AddressResponseDTO»"
    },
...
    "BeerControlResponse«BeerResponseDTO»":{
      "type":"object",
      "properties":{
        "data":{
          "$ref":"#/definitions/BeerResponseDTO"
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«BeerResponseDTO»"
    },
    "BeerControlResponse«List«AddressResponseDTO»»":{
      "type":"object",
      "properties":{
        "data":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/AddressResponseDTO"
          }
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«List«AddressResponseDTO»»"
    },
...
    "BeerControlResponse«List«PersonResponseDTO»»":{
      "type":"object",
      "properties":{
        "data":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/PersonResponseDTO"
          }
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«List«PersonResponseDTO»»"
    },
    "BeerControlResponse«List«PhoneResponseDTO»»":{
      "type":"object",
      "properties":{
        "data":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/PhoneResponseDTO"
          }
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«List«PhoneResponseDTO»»"
    },
...
    "BeerControlResponse«PersonResponseDTO»":{
      "type":"object",
      "properties":{
        "data":{
          "$ref":"#/definitions/PersonResponseDTO"
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«PersonResponseDTO»"
    },
    "BeerControlResponse«PhoneResponseDTO»":{
      "type":"object",
      "properties":{
        "data":{
          "$ref":"#/definitions/PhoneResponseDTO"
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«PhoneResponseDTO»"
    },
    "BeerControlResponse«WarehouseResponseDTO»":{
      "type":"object",
      "properties":{
        "data":{
          "$ref":"#/definitions/WarehouseResponseDTO"
        },
        "duration":{
          "type":"string"
        },
        "pagina":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"BeerControlResponse«WarehouseResponseDTO»"
    },
...
    "BeerResponseDTO":{
      "type":"object",
      "properties":{
        "beerType":{
          "type":"string",
          "enum":[
            "ALE",
            "IPA",
            "LAGER",
            "MALZBIER",
            "STOUT",
            "WEISS",
            "WITBIER"
          ]
        },
        "brand":{
          "type":"string"
        },
        "created":{
          "type":"string",
          "format":"date-time"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "identifier":{
          "type":"string"
        },
        "name":{
          "type":"string"
        },
        "photo":{
          "type":"string"
        }
      },
      "title":"BeerResponseDTO"
    },
...
    "PersonAddressDTO":{
      "type":"object",
      "properties":{
        "addressName":{
          "type":"string"
        },
        "block":{
          "type":"string"
        },
        "city":{
          "type":"string"
        },
        "complement":{
          "type":"string"
        },
        "country":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "number":{
          "type":"integer",
          "format":"int32"
        },
        "reference":{
          "type":"string"
        },
        "state":{
          "type":"string"
        },
        "street":{
          "type":"string"
        },
        "zipCode":{
          "type":"string"
        }
      },
      "title":"PersonAddressDTO"
    },
...
    "PersonPhoneDTO":{
      "type":"object",
      "properties":{
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "number":{
          "type":"string"
        },
        "prefixInternattional":{
          "type":"integer",
          "format":"int32"
        },
        "prefixNational":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"PersonPhoneDTO"
    },
...
    "PersonRequestDTO":{
      "type":"object",
      "properties":{
        "addresses":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/PersonAddressDTO"
          }
        },
        "birthday":{
          "type":"string",
          "format":"date"
        },
        "cpf":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        },
        "phones":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/PersonPhoneDTO"
          }
        },
        "sex":{
          "type":"string",
          "enum":[
            "ASEXUAL",
            "BISEXUAL",
            "FEMALE",
            "HOMOSEXUAL",
            "MALE",
            "PANSEXUAL"
          ]
        },
        "surname":{
          "type":"string"
        }
      },
      "title":"PersonRequestDTO"
    },
    "PersonResponseDTO":{
      "type":"object",
      "properties":{
        "addresses":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/PersonAddressDTO0"
          }
        },
        "birthday":{
          "type":"string",
          "format":"date"
        },
        "cpf":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        },
        "phones":{
          "type":"array",
          "items":{
            "$ref":"#/definitions/PersonPhoneDTO0"
          }
        },
        "sex":{
          "type":"string",
          "enum":[
            "ASEXUAL",
            "BISEXUAL",
            "FEMALE",
            "HOMOSEXUAL",
            "MALE",
            "PANSEXUAL"
          ]
        },
        "surname":{
          "type":"string"
        }
      },
      "title":"PersonResponseDTO"
    },
...
    "PhonePersonDTO0":{
      "type":"object",
      "properties":{
        "birthday":{
          "type":"string",
          "format":"date"
        },
        "cpf":{
          "type":"string"
        },
        "description":{
          "type":"string"
        },
        "email":{
          "type":"string"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "name":{
          "type":"string"
        },
        "sex":{
          "type":"string",
          "enum":[
            "ASEXUAL",
            "BISEXUAL",
            "FEMALE",
            "HOMOSEXUAL",
            "MALE",
            "PANSEXUAL"
          ]
        },
        "surname":{
          "type":"string"
        }
      },
      "title":"PhonePersonDTO0"
    },
    "PhoneRequestDTO":{
      "type":"object",
      "properties":{
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "number":{
          "type":"string"
        },
        "person":{
          "$ref":"#/definitions/PhonePersonDTO0"
        },
        "prefixInternattional":{
          "type":"integer",
          "format":"int32"
        },
        "prefixNational":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"PhoneRequestDTO"
    },
    "PhoneResponseDTO":{
      "type":"object",
      "properties":{
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "number":{
          "type":"string"
        },
        "person":{
          "$ref":"#/definitions/PhonePersonDTO"
        },
        "prefixInternattional":{
          "type":"integer",
          "format":"int32"
        },
        "prefixNational":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"PhoneResponseDTO"
    },
...
    "WarehouseResponseDTO":{
      "type":"object",
      "properties":{
        "beer":{
          "$ref":"#/definitions/BeerResponseDTO"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "quantity":{
          "type":"integer",
          "format":"int64"
        },
        "quantityMaxExpectd":{
          "type":"integer",
          "format":"int64"
        },
        "quantityMin":{
          "type":"integer",
          "format":"int64"
        },
        "update":{
          "type":"string",
          "format":"date-time"
        }
      },
      "title":"WarehouseResponseDTO"
    }
  }
}